cleans up jwt
