+++ app/controllers/application_controller.rb
@@ -1,8 +1,7 @@
 class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
-  skip_before_action :verify_authenticity_token
-  #protect_from_forgery with: :null_session
+  protect_from_forgery with: :null_session
   before_action :load_schema, unless: -> { Hours.single_tenant_mode? }
   before_action :authenticate_user!
   before_action :configure_permitted_parameters, if: :devise_controller?
